+++ tool/download_sqlite.dart
@@ -5,7 +5,7 @@ import 'dart:io';
 const sqlitePath = 'sqlite-amalgamation-3530300';
 const sqliteSource = 'https://sqlite.org/2026/$sqlitePath.zip';
 const sqliteMultipleCiphersSource =
-    'https://github.com/utelle/SQLite3MultipleCiphers/releases/download/v2.3.5/sqlite3mc-2.3.5-sqlite-3.53.2-amalgamation.zip';
+    'https://github.com/utelle/SQLite3MultipleCiphers/releases/download/v2.3.6/sqlite3mc-2.3.6-sqlite-3.53.3-amalgamation.zip';
 
 const sqlcipherVersion = '4.17.0';
 const sqlcipherSource =