+++ sqlite3/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Update SQLite to 3.53.3 and SQLCipher to 4.17.0.
 - Mark a few functions as `isLeaf`, reducing the native overhead of calling them.
 - Reset statements in `execute` and `select`.
+- Allow custom download URLs for pre-compiled binaries through the `url_pattern` user-define.
 
 ## 3.3.4